Refactored the controller code.  Added proper dependency loading in init.rb.
