+++ app/views/project_emails/_edit.html.erb
@@ -1,16 +1,17 @@
-<% form_for @project, :url => {:controller => 'project_emails',
-     :action => 'update', :id => @project} do |f| %>
+<% form_tag project_project_email_path(@project), :method => :put do |f| %>
 <div class="box tabular">
   <p>
-    <%= f.label :email %>
-    <%= f.text_field :email, :value => (params[:email] || @project.email) %>
+    <%= label_tag :email %>
+    <%= text_field_tag :email, params[:email] || @project.email %>
   </p>
 </div>
 <%= submit_tag l(:button_update) %>
 <% end %>
 
-<div class="contextual">
-  <%= link_to l(:button_reset), {:controller => 'project_emails',
-        :action => 'restore_default', :id => @project},
-      :class => "icon icon-del" %>
-</div>
+<% if @project.project_email %>
+  <div class="contextual">
+    <%= link_to l(:button_reset), project_project_email_path(@project),
+        :method => :delete, :confirm => l(:text_are_you_sure),
+        :class => "icon icon-del" %>
+  </div>
+<% end %>
