+++ lib/redmine_project_specific_email_sender/project_patch.rb
@@ -1,0 +1,22 @@
+module RedmineProjectSpecificEmailSender
+  module ProjectPatch
+    
+    def self.included(base)
+      base.send(:include, InstanceMethods)
+      base.class_eval do
+        has_one :project_email
+      end
+    end
+    
+    module InstanceMethods
+      def email
+        project_email ? project_email.email : Setting.mail_from
+      end
+      
+      def email=(email_address)
+        project_email ? (project_email.email = email_address) : build_project_email(:email => email_address)
+      end
+    end
+    
+  end
+end
